Fix #328 acl auth in spark 3 and update jedis, spark version

+++ src/main/scala/com/redislabs/provider/redis/redisFunctions.scala
@@ -401,7 +401,7 @@ object RedisContext extends Serializable {
           pipeline.set(k, v)
         }
         else {
-          pipeline.setex(k, ttl, v)
+          pipeline.setex(k, ttl.toLong, v)
         }
       }
       conn.close()
@@ -422,7 +422,7 @@ object RedisContext extends Serializable {
     val pipeline = foreachWithPipelineNoLastSync(conn, arr) { case (pipeline, (k, v)) =>
       pipeline.hset(hashName, k, v)
     }
-    if (ttl > 0) pipeline.expire(hashName, ttl)
+    if (ttl > 0) pipeline.expire(hashName, ttl.toLong)
     pipeline.sync()
     conn.close()
   }
@@ -448,7 +448,7 @@ object RedisContext extends Serializable {
           foreachWithPipeline(conn, arr) { (pipeline, a) =>
             val (key, hashFields) = a._2
             pipeline.hmset(key, hashFields)
-            if (ttl > 0) pipeline.expire(key, ttl)
+            if (ttl > 0) pipeline.expire(key, ttl.toLong)
           }
         }
       }
@@ -478,7 +478,7 @@ object RedisContext extends Serializable {
               foreachWithPipeline(conn, arr) { (pipeline, a) =>
                 val (key, hashFields) = a._2
                 pipeline.hmset(key, hashFields)
-                if (ttl > 0) pipeline.expire(key, ttl)
+                if (ttl > 0) pipeline.expire(key, ttl.toLong)
               }
             }
           }
@@ -498,7 +498,7 @@ object RedisContext extends Serializable {
     val pipeline = foreachWithPipelineNoLastSync(conn, arr) { case (pipeline, (k, v)) =>
       pipeline.zadd(zsetName, v.toDouble, k)
     }
-    if (ttl > 0) pipeline.expire(zsetName, ttl)
+    if (ttl > 0) pipeline.expire(zsetName, ttl.toLong)
     pipeline.sync()
     conn.close()
   }
@@ -516,7 +516,7 @@ object RedisContext extends Serializable {
     val pipeline = foreachWithPipelineNoLastSync(conn, arr) { (pipeline, v) =>
       pipeline.sadd(setName, v)
     }
-    if (ttl > 0) pipeline.expire(setName, ttl)
+    if (ttl > 0) pipeline.expire(setName, ttl.toLong)
     pipeline.sync()
     conn.close()
   }
@@ -537,7 +537,7 @@ object RedisContext extends Serializable {
     val pipeline = foreachWithPipelineNoLastSync(conn, arr) { (pipeline, v) =>
       pipeline.rpush(listName, v)
     }
-    if (ttl > 0) pipeline.expire(listName, ttl)
+    if (ttl > 0) pipeline.expire(listName, ttl.toLong)
     pipeline.sync()
     conn.close()
   }
@@ -560,7 +560,7 @@ object RedisContext extends Serializable {
           foreachWithPipeline(conn, arr) { (pipeline, a) =>
             val (key, listVals) = a._2
             pipeline.rpush(key, listVals: _*)
-            if (ttl > 0) pipeline.expire(key, ttl)
+            if (ttl > 0) pipeline.expire(key, ttl.toLong)
           }
         }
       }
@@ -583,7 +583,7 @@ object RedisContext extends Serializable {
           foreachWithPipeline(conn, arr) { (pipeline, a) =>
             val (key, listVals) = a._2
             pipeline.rpush(key, listVals: _*)
-            if (ttl > 0) pipeline.expire(key, ttl)
+            if (ttl > 0) pipeline.expire(key, ttl.toLong)
           }
         }
       }
